move to enterprise_app_installation.go, delete enterprise_apps.go
