+++ github/enterprise_app_installation.go
@@ -157,3 +157,112 @@ func (s *EnterpriseService) UninstallApp(ctx context.Context, enterprise, org st
 
 	return resp, nil
 }
+
+// AppInstallationRepositoriesRequest specifies the parameters for
+// EnterpriseService.AddRepositoriesToAppInstallation and
+// EnterpriseService.RemoveRepositoriesFromAppInstallation.
+type AppInstallationRepositoriesRequest struct {
+	// Repository names to add to or remove from the installation.
+	Repositories []string `json:"repositories"`
+}
+
+// UpdateAppInstallationRepositoriesRequest specifies the parameters for
+// EnterpriseService.UpdateAppInstallationRepositories.
+type UpdateAppInstallationRepositoriesRequest struct {
+	// Can be "all" or "selected".
+	RepositorySelection *string `json:"repository_selection,omitempty"`
+	// Repository names to grant the installation access to. Only required
+	// when RepositorySelection is "selected".
+	Repositories []string `json:"repositories,omitempty"`
+}
+
+// ListRepositoriesForOrgAppInstallation lists the repositories that an enterprise app installation
+// has access to on an organization.
+//
+// GitHub API docs: https://docs.github.com/enterprise-cloud@latest/rest/enterprise-admin/organization-installations?apiVersion=2022-11-28#get-the-repositories-accessible-to-a-given-github-app-installation
+//
+//meta:operation GET /enterprises/{enterprise}/apps/organizations/{org}/installations/{installation_id}/repositories
+func (s *EnterpriseService) ListRepositoriesForOrgAppInstallation(ctx context.Context, enterprise, org string, installationID int64, opts *ListOptions) ([]*AccessibleRepository, *Response, error) {
+	u := fmt.Sprintf("enterprises/%v/apps/organizations/%v/installations/%v/repositories", enterprise, org, installationID)
+	u, err := addOptions(u, opts)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	req, err := s.client.NewRequest(ctx, "GET", u, nil)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	var r []*AccessibleRepository
+	resp, err := s.client.Do(req, &r)
+	if err != nil {
+		return nil, resp, err
+	}
+
+	return r, resp, nil
+}
+
+// UpdateAppInstallationRepositories changes a GitHub App installation's repository access
+// between all repositories and a selected set.
+//
+// GitHub API docs: https://docs.github.com/enterprise-cloud@latest/rest/enterprise-admin/organization-installations?apiVersion=2022-11-28#toggle-installation-repository-access-between-selected-and-all-repositories
+//
+//meta:operation PATCH /enterprises/{enterprise}/apps/organizations/{org}/installations/{installation_id}/repositories
+func (s *EnterpriseService) UpdateAppInstallationRepositories(ctx context.Context, enterprise, org string, installationID int64, opts UpdateAppInstallationRepositoriesRequest) (*Installation, *Response, error) {
+	u := fmt.Sprintf("enterprises/%v/apps/organizations/%v/installations/%v/repositories", enterprise, org, installationID)
+	req, err := s.client.NewRequest(ctx, "PATCH", u, opts)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	var r *Installation
+	resp, err := s.client.Do(req, &r)
+	if err != nil {
+		return nil, resp, err
+	}
+
+	return r, resp, nil
+}
+
+// AddRepositoriesToAppInstallation grants repository access for a GitHub App installation.
+//
+// GitHub API docs: https://docs.github.com/enterprise-cloud@latest/rest/enterprise-admin/organization-installations?apiVersion=2022-11-28#grant-repository-access-to-an-organization-installation
+//
+//meta:operation PATCH /enterprises/{enterprise}/apps/organizations/{org}/installations/{installation_id}/repositories/add
+func (s *EnterpriseService) AddRepositoriesToAppInstallation(ctx context.Context, enterprise, org string, installationID int64, opts AppInstallationRepositoriesRequest) ([]*AccessibleRepository, *Response, error) {
+	u := fmt.Sprintf("enterprises/%v/apps/organizations/%v/installations/%v/repositories/add", enterprise, org, installationID)
+	req, err := s.client.NewRequest(ctx, "PATCH", u, opts)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	var r []*AccessibleRepository
+	resp, err := s.client.Do(req, &r)
+	if err != nil {
+		return nil, resp, err
+	}
+
+	return r, resp, nil
+}
+
+// RemoveRepositoriesFromAppInstallation revokes repository access from a GitHub App installation.
+//
+// GitHub API docs: https://docs.github.com/enterprise-cloud@latest/rest/enterprise-admin/organization-installations?apiVersion=2022-11-28#remove-repository-access-from-an-organization-installation
+//
+//meta:operation PATCH /enterprises/{enterprise}/apps/organizations/{org}/installations/{installation_id}/repositories/remove
+func (s *EnterpriseService) RemoveRepositoriesFromAppInstallation(ctx context.Context, enterprise, org string, installationID int64, opts AppInstallationRepositoriesRequest) ([]*AccessibleRepository, *Response, error) {
+	u := fmt.Sprintf("enterprises/%v/apps/organizations/%v/installations/%v/repositories/remove", enterprise, org, installationID)
+	req, err := s.client.NewRequest(ctx, "PATCH", u, opts)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	var r []*AccessibleRepository
+	resp, err := s.client.Do(req, &r)
+	if err != nil {
+		return nil, resp, err
+	}
+
+	return r, resp, nil
+}
